requirements.txt
Flask
torch
torchvision
Pillow
werkzeug
